requirements.txt
apache-airflow==2.9.2
requests
pandas
azure-storage-blob
python-dotenv
apache-airflow==2.9.2
apache-airflow-providers-databricks==6.0.0
apache-airflow-providers-microsoft-azure
pandas
requests
